host/requirements.txt
aioclient==0.2.2
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
boto3==1.35.64
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
connect-python==0.5.0
fastapi==0.121.0
python-dotenv==1.0.0
frozenlist==1.8.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
propcache==0.4.1
protobuf==6.33.0
pycares==4.11.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
